+++ R/plotLoadings.pca.R
@@ -72,9 +72,10 @@ xlim = NULL,
     df = data.frame(importance = value.selected.var) # contribution of the loading
     
     # barplot with contributions
+    colnames.X <- .trim_long_names(colnames.X) ## issue 45
     par(mar = c(4, max(7, max(sapply(colnames.X, nchar),na.rm = TRUE)/3), 4, 2))
 
-    mp = barplot(df$importance, horiz = TRUE, las = 1, col = col, axisnames = TRUE, names.arg = colnames.X, #names.arg = row.names(df),
+    barplot(df$importance, horiz = TRUE, las = 1, col = col, axisnames = TRUE, names.arg = colnames.X, #names.arg = row.names(df),
     cex.names = size.name, cex.axis = 0.7, beside = TRUE, border = border, xlim = xlim)
     
     if (is.null(title))