+++ R/plotLoadings.splsda.R
@@ -154,6 +154,7 @@ xlim = NULL,
             }
 
             # display barplot with names of variables
+            colnames.X <- .trim_long_names(colnames.X) ## issue 45
             if (plot) # condition if all we need is the contribution stats
             {
                 if (!is.null(title) & length(block) > 1)
@@ -163,7 +164,7 @@ xlim = NULL,
                     par(mar = c(4, max(7, max(sapply(colnames.X, nchar), na.rm = TRUE)/3), 4, 2))
                 }
                 
-                mp = barplot(df$importance, horiz = TRUE, las = 1, col = df$color, axisnames = TRUE, names.arg = colnames.X, #names.arg = row.names(df),
+                barplot(df$importance, horiz = TRUE, las = 1, col = df$color, axisnames = TRUE, names.arg = colnames.X, #names.arg = row.names(df),
                 cex.names = size.name, cex.axis = 0.7, beside = TRUE, border = border, xlim = xlim[i, ])
                 
                 if ( length(block) == 1 & is.null(title) )
@@ -181,7 +182,7 @@ xlim = NULL,
                 {
                     par(mar = c(5, 0, 4, 3) + 0.1)
                     plot(1,1, type = "n", axes = FALSE, ann = FALSE)
-                    legend(0.8, 1, col = legend.color[1:nlevels(Y)], legend = levels(Y), pch = 19,
+                    legend(0.8, 1.1, col = legend.color[1:nlevels(Y)], legend = levels(Y), pch = 19,
                     title = paste(legend.title),
                     cex = size.legend)
                 }